refactor: use Decimal for kWh to prevent float accumulation error
Summing hundreds of IEEE 754 floats drifts noticeably at monthly and
annual aggregation granularity. Decimal arithmetic is exact for the
3-decimal-place values the API returns.

- ConsumptionRecord.kwh: float → Decimal
- fetch.py: Decimal(str(raw)) converts the JSON-parsed value without
  passing through a float intermediate
- cli.py: float(record.kwh) for JSON serialisation only — individual
  values are fine as float; precision matters in aggregation
- All test RECORD constants and assertions updated to Decimal

+++ src/oomi_influx/models.py
@@ -1,11 +1,12 @@
 from dataclasses import dataclass
 from datetime import datetime
+from decimal import Decimal
 
 
 @dataclass
 class ConsumptionRecord:
     timestamp: datetime
-    kwh: float
+    kwh: Decimal
 
 
 class LoginError(Exception):
